requirements.txt
langchain
langchain-community
sentence-transformers
chromadb
python-dotenv
requests
uvicorn
